+++ crates/db/src/impls/comment/comment_votes.rs
@@ -14,7 +14,7 @@ impl CommentVote {
         pool: &DbPool,
         ids: Vec<i32>,
         for_user_id: i32,
-    ) -> Result<Vec<(i32, i16)>, Error> {
+    ) -> Result<Vec<(i32, i32)>, Error> {
         let conn = &mut get_conn(pool).await?;
         use crate::schema::{comment_votes, comments};
 
@@ -26,12 +26,12 @@ impl CommentVote {
             )
             .filter(comments::id.eq_any(ids))
             .select((comments::id, comment_votes::score.nullable()))
-            .load::<(i32, Option<i16>)>(conn)
+            .load::<(i32, Option<i32>)>(conn)
             .await
             .map(|list| {
                 list.into_iter()
                     .map(|(comment_id, vote_type)| (comment_id, vote_type.unwrap_or(0)))
-                    .collect::<Vec<(i32, i16)>>()
+                    .collect::<Vec<(i32, i32)>>()
             })
     }
 }
